Markov model code
overChange.py calculates the total number of runs made by a batting order. For simplicity, we check only three orders.

If you have time and the resources, you can check out all the 11! possibilities.

The final_scores file gives the final scores of the 'one-man teams' and this is used to rank the players.
(Pls dunt)